android: fix likely minZoom/maxZoom compile error in offline map download
MapboxOfflineRegion.kt is this app's own flagged highest-risk file for a
real compile error (never built, "most likely to need small signature
fixes" per its own doc). Verified against Mapbox's real, current
createTilesetDescriptor docs (v11.4.0) via web search rather than guessing:
the actual documented sample uses plain Int literals (.minZoom(0)
.maxZoom(16)), not Byte. Kotlin has no implicit Byte<->Int widening, so the
previous .minZoom(MIN_ZOOM.toInt().toByte()) call was very likely a genuine
compile error waiting to be hit the first time this file actually builds
against the real SDK jar. Fixed to match the documented real usage.

Not yet compiled against the real SDK -- Ryzen is on a different task
right now -- so this is evidence-based, not compiler-confirmed. Left a
comment pointing back to Mapbox's current offline guide if this specific
fix turns out to still be wrong (the SDK's shape has moved before).

+++ android/app/src/main/java/au/com/threesixty/cabdispatch/data/remote/MapboxOfflineRegion.kt
@@ -106,8 +106,17 @@ object MapboxOfflineRegion {
         val tilesetDescriptor = offlineManager.createTilesetDescriptor(
             TilesetDescriptorOptions.Builder()
                 .styleURI(Style.DARK)
-                .minZoom(MIN_ZOOM.toInt().toByte())
-                .maxZoom(MAX_ZOOM.toInt().toByte())
+                // Plain Int, NOT .toByte() -- verified 2026-08-27 against Mapbox's real
+                // v11.4.0 createTilesetDescriptor docs sample (.minZoom(0).maxZoom(16), no
+                // Byte conversion anywhere). Kotlin does not implicitly widen Byte<->Int, so
+                // the previous .toByte() call was very likely a genuine compile error the
+                // moment this file was first built against the real SDK jar -- fixed before
+                // that ever needed to be discovered the hard way. If this still doesn't
+                // compile, minZoom/maxZoom's real signature has moved again since this doc
+                // was checked -- re-verify against docs.mapbox.com's CURRENT offline guide,
+                // don't just guess a different numeric type.
+                .minZoom(MIN_ZOOM.toInt())
+                .maxZoom(MAX_ZOOM.toInt())
                 .build(),
         )
